requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-groq
langchain-huggingface
sentence-transformers
faiss-cpu
numpy
python-dotenv
requests
